refactor: move logView folder to logger
* Relocated LogViewProvider component to be with related logger files
* Updated import paths in extension.ts and logUtils.ts to reflect the new structure
