+++ coauthor-for-vs-code/src/logger/LogViewProvider.ts
